add is_shift to RTC process workflow
