+++ hyp3_gamma/rtc/rtc_sentinel.py
@@ -35,7 +35,7 @@ import hyp3_gamma
 from hyp3_gamma.dem import get_geometry_from_kml, prepare_dem_geotiff
 from hyp3_gamma.metadata import create_metadata_file_set_rtc
 from hyp3_gamma.rtc.coregistration import CoregistrationError, check_coregistration
-from hyp3_gamma.util import set_pixel_as_point, unzip_granule
+from hyp3_gamma.util import set_pixel_as_point, unzip_granule, is_shift
 
 
 log = logging.getLogger()
@@ -405,8 +405,11 @@ def rtc_sentinel_gamma(safe_dir: str, resolution: float = 30.0, radiometry: str 
         if not include_rgb:
             os.remove(rgb_tif)
 
-    for tif_file in glob(f'{product_name}/*.tif'):
-        set_pixel_as_point(tif_file, shift_origin=dem_name == 'legacy')
+    # do pixel shift if needed
+    if is_shift(mli_par, 'dem_seg.par', power_tif)[0]:
+        for tif_file in glob(f'{product_name}/*.tif'):
+            set_pixel_as_point(tif_file, shift_origin=True)
+
     cogify_dir(directory=product_name)
 
     log.info('Generating browse images and metadata files')
